Added Exit Button
Well, the button  _was_ there, but in this commit I added functionality
to it. Now that I pretty much finished the main menu, I shall add two
branches to the project, Singleplayer and Multiplayer, so that I can
work on scripting the multiplayer and Raf can work on singleplayer
without any conflicts.